pyproject.toml
[project]
name = "neo4j-spacy-procedures"
version = "0.1.0"
description = "neo4j spacy integration"
readme = "README.md"
requires-python = ">=3.11.10"
dependencies = [
    "en-core-web-sm",
    "neo4j-driver>=5.26.0",
    "spacy>=3.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
select = ["E", "F", "I", "N", "W"]
line-length = 88

[tool.hatch.build.targets.wheel]
packages = ["src/neo4j_spacy_procedures"]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.2.0",
]
